requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
cffi==2.0.0
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
fastapi==0.136.0
h11==0.16.0
idna==3.11
pycparser==3.0
pydantic==2.13.1
pydantic_core==2.46.1
PyJWT==2.12.1
python-dotenv==1.0.1
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
